+++ src/types/signer.ts
@@ -1,5 +1,7 @@
+import { StarknetChainId } from '../constants';
 import { InvocationsDetails } from './lib';
 
 export interface InvocationsSignerDetails extends Required<InvocationsDetails> {
   walletAddress: string;
+  chainId: StarknetChainId;
 }